pose-detection-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.0
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
pymongo==4.6.0
gunicorn==21.2.0
ultralytics==8.0.196
scipy==1.11.4
requests==2.31.0

